bisync: improve handling of nested empty dirs
Before this change, when nested dirs were deleted from one path, bisync (with --create-empty-src-dirs specified) would remove the corresponding dirs from the other path in ascending alphabetical order. For example, bisync would first Rmdirs 'remote:a', and then attempt to Rmdirs 'remote:a/b'. This would result in a (non-critical) error, because 'remote:a/b' was already removed as part of step 1, and thus no longer existed by step 2.

After this change, the list of empty dirs to remove is sorted in descending order, so that any subdirs should be removed before attempting to remove their parents. This eliminates the "directory not found" errors and reduces unnecessary operations.