bisync: add basic GUI status dashboard
This introduces a basic status dashboard GUI for bisync (disabled by default;
`--gui` to enable). It is essentially a local HTML webpage that shows a table
of your recent bisync runs, along with their status and some stats and other
info. The main goal of the GUI is to make bisync a little friendlier to use
headlessly and see at a quick glance whether anything needs your attention,
without having to go diving through log files.

In this first version, the GUI is just a static read-only HTML page that bisync
writes to at various points during a bisync run. In future versions it may
become more interactive. Crucially, since it's just a static local file, there
is no "server" or other networking required -- you can check it entirely
offline, even when rclone is not running!

It should load in any modern internet browser, and the URL remains constant for
easy bookmarking. By default it lives in the `--workdir`, but you can set a
custom directory with `--gui-dir`. The `--gui-dir` can even be another rclone
remote (for example, you could keep it in Drive/Dropbox for checking
on-the-go), but it is recommended to keep it on `local` for the best
performance.

Several parameters can be customized -- see the [command
flags](#command-line-syntax) starting with `--gui`.

Examples:
docs/static/img/bisync_status_example.html
docs/static/img/bisync_gui_example.png