local: preserve macOS metadata on non-local remotes
Before this change, macOS-specific metadata was not preserved by rclone, even for
local-to-local transfers (it does not use the "user." prefix, nor is Mac metadata
limited to xattrs.)

After this change, macOS-specific metadata can be preserved by storing it as an
rclone metadata value in "AppleDouble" format (only when the source is local on
macOS and --local-apple-double is set). In the reverse direction (dest is macOS
local) rclone will automatically "unpack" this metadata to its original form
(for instance, as mac xattrs).

Note that these values can sometimes exceed the length limitations imposed by
some backends (such as Drive) on metadata values. For this reason, an additional --
local-metadata-max-length flag has been added to allow skipping values that would
otherwise error.

Using the "AppleDouble" format in this way has an advantage over using Go's
xattr package, as AppleDouble is the native and idiomatic format for storing
macOS metadata and can preserve not just xattrs but also metadata such as Finder
and Spotlight params. It is also (arguably) preferable to rsync's approach,
which requires a separate ._prefixed sidecar for every file (hidden on macOS but
often visible and messy on other remotes, and which also runs the risk of
resource forks becoming separated from their data forks.)