bisync: fix data race on local tests
Before this change, TestBisyncConcurrent could error if one of the parallel
tests finished while the other was in the middle of a bisync run.

The reason is that when the subtest completes, it calls the cleanup function
returned by fstest.RandomRemote(). This cleanup function attempts to walk.ListR
the RandomRemoteName (which we previously ignored on local), and errors because
it doesn't exist. Because this error is counted in the global stats, the other
parallel subtest thinks there has been a bisync error (even though there hasn't
been).

This change fixes it by using the RandomRemoteName on local, when we have it.
Because we're using it, it should exist when the cleanup function tries to walk
and purge it, and thus should not error.